Remove hardcoded values and make solution a bit more secure
Jenkins config moved to a separate file that is not part of the repo and supposed to be
provieded as build input.

We will still have password as a plain text inside startup script which is a security threat
but you need to get into the image to get it. Because of that I consider this solution as good enough

Signed-off-by: Andrey Devyatkin <andrey.a.devyatkin@gmail.com>